fix: change dependencies load strcuture
