+++ src/program.ts
@@ -1,7 +1,11 @@
 #!/usr/bin/env node
 import { program } from 'commander';
 
-import getNodeLicense from './node';
+import getNodeLicenses from './node';
+import getGradleLicenses from './gradle';
+import outputMessage from './output';
+import resultToHTML from './output/resultToHTML';
+import resultToJson from './output/resultToJson';
 
 const packageJson = require(`${__dirname}/../package.json`);
 
@@ -12,5 +16,32 @@ program.option('--template [templateFile]', 'Custom template for export to html 
 program.parse(process.argv);
 
 const options = program.opts();
+const launch = async (opts: any) => {
+  const { json, html, template } = opts;
 
-getNodeLicense(options.json, options.html, options.template);
+  const nodeLicenses = getNodeLicenses();
+  const gradleLicenses = await getGradleLicenses();
+  const licenses = [...nodeLicenses, ...gradleLicenses];
+
+  console.log(licenses);
+
+  if (typeof json === 'boolean' || typeof json === 'string') {
+    outputMessage('cyan', 'Result export to json..');
+
+    const result = resultToJson(licenses, typeof json === 'string' ? json : null);
+
+    outputMessage('orange', `Result is saved at ${result}`);
+  }
+
+  if (typeof html === 'boolean' || typeof html === 'string') {
+    outputMessage('cyan', 'Result export to html..');
+
+    const result = typeof template === 'string'
+      ? resultToHTML(licenses, typeof html === 'string' ? html : null, template)
+      : resultToHTML(licenses, typeof html === 'string' ? html : null);
+
+    outputMessage('orange', `Result is saved at ${result}`);
+  }
+};
+
+launch(options);
